sp_api/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
python-jose==3.3.0
pandas
tabulate==0.9.0
xlsxwriter==3.1.2
pyodbc
aiohttp==3.9.1